Move helper functions in proper utils modules
